requirements.txt
python-telegram-bot
feedparser
requests
python-dotenv
